+++ je_auto_control/utils/selection_view/__init__.py
@@ -1,0 +1,6 @@
+"""Container selection state + view switching (Selection / MultipleView patterns)."""
+from je_auto_control.utils.selection_view.selection_view import (
+    get_selection, list_views, set_view,
+)
+
+__all__ = ["get_selection", "list_views", "set_view"]